requirements.txt
adbc_driver_postgresql==1.1.0
psycopg2-binary==2.9.9
pyarrow==17.0.0
pydantic[dotenv]==2.8.2
pydantic-settings==2.4.0
typer==0.12.4